refactor(ui): extract progress card to component
