+++ web/src/routes/(site)/+layout.svelte
@@ -5,7 +5,7 @@
 </script>
 
 <main
-	class="mx-auto flex min-h-[100vh] max-w-[1600px] flex-col items-center justify-center overflow-auto px-8"
+	class="mx-auto flex min-h-[100vh] max-w-[1600px] flex-col items-center overflow-auto py-6 px-8"
 >
 	{@render children()}
 </main>
