+++ web/src/routes/(site)/calendar/[date]/+page.svelte
@@ -23,6 +23,8 @@
 		CardTitle
 	} from '$lib/components/ui/card';
 	import { Badge } from '$lib/components/ui/badge';
+	import { ArrowLeft } from "@lucide/svelte";
+	import ProgressCard from './ProgressCard.svelte';
 
 	let { data }: PageProps = $props();
 
@@ -60,89 +62,29 @@
 		};
 	});
 
+	const dateFormatter = new DateFormatter('da-DK', { dateStyle: 'long' });
+
+	const formattedDate = $derived(
+		dateFormatter.format(
+			new Date(daySummary.date)
+		)
+	)
+
 	const categoryTriggerContent = $derived(
 		data.categories.find((c) => c.id == $form.categoryId)?.title ?? 'Vælg en kategori'
 	);
-
-	const remainingHours = $derived(
-		Math.max((daySummary.maxHours - daySummary.totalHours) / Hour, 0)
-	);
-
-	const progressPercentage = $derived(
-		Math.min((daySummary.totalHours / daySummary.maxHours) * 100, 100)
-	);
-
-	const getProgressColor = () => {
-		if (progressPercentage < 50) return 'stroke-blue-500';
-		if (progressPercentage < 75) return 'stroke-yellow-500';
-		if (progressPercentage < 100) return 'stroke-orange-500';
-		return 'stroke-green-500';
-	};
 </script>
 
-<div class="grid w-full grid-cols-2 gap-2 p-2">
-	<Card>
-		<CardHeader class="text-center">
-			<CardTitle class="text-2xl">Dagens opsummering</CardTitle>
-		</CardHeader>
-		<CardContent class="flex h-full flex-col items-center gap-8">
-			<div class="relative h-48 w-48">
-				<svg class="h-48 w-48 -rotate-90 transform" viewBox="0 0 100 100">
-					<circle
-						cx="50"
-						cy="50"
-						r="45"
-						stroke="currentColor"
-						stroke-width="8"
-						fill="transparent"
-						class="text-slate-200"
-					/>
-					<circle
-						cx="50"
-						cy="50"
-						r="45"
-						stroke="currentColor"
-						stroke-width="8"
-						fill="transparent"
-						stroke-dasharray={`${2 * Math.PI * 45}`}
-						stroke-dashoffset={`${2 * Math.PI * 45 * (1 - progressPercentage / 100)}`}
-						class={`${getProgressColor()} transition-all duration-1000 ease-out`}
-						stroke-linecap="round"
-					/>
-				</svg>
+<div class="flex w-full items-center p-4 border rounded-xl mb-6 gap-4">
+	<Button href="/calendar">
+		<ArrowLeft />
+		Tilbage
+	</Button>
+	<p class="font-semibold text-xl tracking-tight">Tidsregistrering d. {formattedDate}</p>
+</div>
 
-				<div class="absolute inset-0 flex flex-col items-center justify-center">
-					<div class="text-3xl font-bold text-slate-800">
-						{(daySummary.totalHours / Hour).toFixed(1)}t
-					</div>
-					<div class="text-sm text-slate-500">
-						af {(daySummary.maxHours / Hour).toFixed(1)} timer
-					</div>
-					{#if remainingHours > 0}
-						<div class="mt-1 text-xs text-slate-400">Mangler {remainingHours.toFixed(1)}t</div>
-					{/if}
-				</div>
-			</div>
-			<div class="grid w-full grid-cols-3">
-				<div class="flex flex-col items-center justify-center">
-					<p class="text-lg font-semibold">{daySummary.timeEntries.length}</p>
-					<p class="text-muted-foreground text-sm">Registreringer</p>
-				</div>
-				<div class="flex flex-col items-center justify-center">
-					<p class="text-lg font-semibold">
-						{((daySummary.totalHours / daySummary.maxHours) * 100).toFixed(1)}%
-					</p>
-					<p class="text-muted-foreground text-sm">Udførsel</p>
-				</div>
-				<div class="flex flex-col items-center justify-center">
-					<p class="text-lg font-semibold">
-						{daySummary.totalHours / Hour / daySummary.timeEntries.length}t
-					</p>
-					<p class="text-muted-foreground text-sm">Gns / Registrering</p>
-				</div>
-			</div>
-		</CardContent>
-	</Card>
+<div class="grid w-full grid-cols-2 gap-6">
+	<ProgressCard {daySummary} />
 
 	<Card>
 		<CardContent class="h-full">
@@ -204,9 +146,7 @@
 	<Card class="col-span-2">
 		<CardHeader>
 			<CardTitle>
-				Registreringer d. {new DateFormatter('da-DK', { dateStyle: 'long' }).format(
-					new Date(daySummary.date)
-				)}
+				Registreringer d. {formattedDate}
 			</CardTitle>
 			<CardDescription
 				>Timer: {daySummary.totalHours / Hour} Max timer: {daySummary.maxHours /
